ID:5005 Refactoring of whitelisting mechanism
Putting placeholder into variable, so I do not have to repeat it so
often (arguably the trailing __ should be treated the same way?).
Renaming $ref -> $queryRefType for readability.
Renaming $query -> $queryRef in _filterLocalParams to make clear, that
it is not a string.
Created new sub to replace/unreplace placeholders.
Changed regex to not be fooled by spaces, ie. also match '{ !'.